+++ .github/CHANGELOG.md
@@ -3,6 +3,8 @@
 Update <small>_ March 2024</small>
 
 - feat: adds github issue command (18/03/2024)
+- fix: content change in welcome command (18/03/2024)
+- fix: multiple plus symbols in zulu command (18/03/2024)
 - feat: add infractions counts to buttons in list infractions command (18/03/2024)
 
 Update <small>_ February 2024</small>